+++ engine/src/main/java/net/jqwik/engine/properties/DefaultArbitraries.java
@@ -19,6 +19,7 @@ public class DefaultArbitraries {
 		providers.add(new ListArbitraryProvider());
 		providers.add(new SetArbitraryProvider());
 		providers.add(new MapArbitraryProvider());
+		providers.add(new EntryArbitraryProvider());
 		providers.add(new StreamArbitraryProvider());
 		providers.add(new OptionalArbitraryProvider());
 		providers.add(new ArrayArbitraryProvider());